// src/pages/ArticlesPage/model/types/articlesPageScheme.ts
import { EntityState } from '@reduxjs/toolkit';
import {
	Article,
	ArticleSortField,
	ArticleType,
	ArticleView,
} from 'entities/Article';

import { SortOrder } from 'shared/types';

export interface ArticlesPageScheme extends EntityState<Article, string> {
	isLoading?: boolean;
	error?: string;

	// pagination
	page: number;
	limit: number;
	hasMore: boolean;

	// filter
	view: ArticleView;
	order: SortOrder;
	sort: ArticleSortField;
	search: string;
	type: ArticleType;

	_inited: boolean;
}